requirements.txt
pyyaml
pyobjc-framework-Quartz
keyboard
soco
